backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
reportlab>=4.0.0
python-decouple>=3.8
django-cors-headers>=3.14.03.1
python-decouple==3.8
Pillow==10.1.0
